+++ src/kit/ui/UiTexts.hpp
@@ -16,7 +16,6 @@
 #include "../lib/unilib/utf8.h"
 
 #include "UiTheme.hpp"
-#include "../utils/UtilsTexts.hh"
 
 
 #define TEXTS_DEFAULT_FONT_COLOR (unsigned int) RGBA8(0, 0, 0, 255)
@@ -62,7 +61,11 @@ private:
     UiTheme *theme;
     std::pair<std::string, unsigned int> keyFont;
     std::string family, fontPath;
-    UtilsTexts *utilsTexts;
+
+    int i;
+    std::u32string text32;
+    TextData textDataText;
+    unsigned int posBreak;
 
     void drawFinal(int x, int y, TextStyle textStyle, unsigned int color, bool italic, std::string text);
     void drawFinal(int x, int y, TextStyleData _textStyleData, unsigned int color, std::string text);
@@ -105,6 +108,15 @@ public:
     //Dynamic import
     void cleanFonts();
     void cleanFont(std::string type, int size);
+
+    //Helpers functions
+    int keySearch(const std::string& s, const std::string& key);
+    std::string applyTextWidthLimit(std::string text, int width, TextStyleData textStyleData);
+    std::string applyTextHeightLimit(std::string text, int height, TextStyleData textStyleData);
+
+    //transformation text
+    std::string toUppercase(std::string text);
+    std::string toLowercase(std::string text);
 
 };