+++ src/kit/Changelog.md
@@ -5,7 +5,8 @@
 * Add Scroll class to manage scroll elements
 * Improve RadioBoxes & Checkboxes functions with text to support changes (Text type & color)
 * Add a param to change the size of the button
-* Move applyTextHeightLimit & applyTextWidthLimit of TextFields into UtilsTexts
+* Remove UtilsTexts and revert all changes to UiTexts & UtilsIme to simplify logic
+* Move applyTextHeightLimit & applyTextWidthLimit of TextFields into UiTexts
 * Add possibility to choose a start value in PadTouchController
 * Improve PadTouchController class when the menu change
 * Fixes everywhere, every commit